drfsite/requirements.txt
Django==4.1.5
Django==5.0.3
djangorestframework==3.15.0
pillow==10.2.0